+++ core/src/jsonpp.cpp
@@ -12,7 +12,7 @@
 #include <iomanip>
 #include <limits>
 
-#include "json.hpp"
+#include "jsonpp.hpp"
 
 
 #if defined(_MSC_VER)
