+++ core/src/meson.build
@@ -68,7 +68,7 @@ lib = library('kmnkbp0',
   'km_kbp_state_api.cpp',
   'km_kbp_debug_api.cpp',
   'km_kbp_processevent_api.cpp',
-  'json.cpp',
+  'jsonpp.cpp',
   'mock/mock_processor.cpp',
   'kmx/kmx_consts.cpp',
   'kmx/kmx_processevent.cpp',
