+++ core/tests/unit/json/jsontest.cpp
@@ -14,7 +14,7 @@
 
 #include <iostream>
 #include <fstream>
-#include <json.hpp>
+#include <jsonpp.hpp>
 
 #ifdef __EMSCRIPTEN__
 #include <emscripten.h>
